Add fbDirectHome FBs
Add 2 fbDirectHome function blocks for virtual axes.
They perform a reverse calculation when coupling the slits.
Depending on the physical axes positions it will calculat what the gap/offset is.
This is then set to the virtual axes posisitions.
State machine is also added to initiate the variables function blocks.