Code change to FB_Slit
Changes in state machine in order to enable control of both
physical axes and virtual axes without having to toggle a boolean.
Now the FB_GearInMultiMaster is only enabled when the
FB_DriveVirtual.bEnable bits are set and disabled when it has finished
executing.
This needs to be changed slightly as they are hardcoded to be
M1 and M2.
Also: problem now is the slit blades can be crashed into each other
when controlling the physical axes.