Change state machine for slit control
Change totally state machine for slit control.
Taken from testing in another project.
This algorithm has the physical axes by default uncoupled.
When a command is recieved for the virtual axes the fbGearInMultiMaster is enabled.
It is then disabled after the move is completed.
This allows the phsyical axes to be driven as well without the user needing to worry about
coupling/decoupling axes.
It also reverse calculates the gap/center axes so it shouldn't need to be initialised.