Add extra collision avoiding fuctionality
If a slit blade gets close to the other blade, within a set distance,
the blade speed (using fOverride) will decrease depending on how close
the blades are. So it will reduce from 100% to almost 0% so the blades
cannot be driven into each other.
The second fucntionality which was already there is to disable the feed
in one direction if the blades go past each other which now shouldn't
be possible.