Modify collision check algorithm
Extend "slow down zone" to extend past the opposite axis.
So now instead of just a small zone the velocity is reduced it extends past
the oppositve axis as well.
Removed another check, that checked the actual position for the move inhibit.
This means that the gap won't go under the minimum soft limit if either
axis overshoots.