Add check that resets state machine if either virtual axis is reset
To enable the state machine to reset if there is an error from EPICS.
There is currently a reset bit or one can toggle the enable bit to FB_Slit
but these are not currently accessable from EPICS. And would be better if
they didn't need to be. So if the Gap of Center recieves a bReset then
it will also reset the statemachine.