Add homing mode 16 and 17
Homing to the zero (index) pulse of an incremental encoder.
16 from positive direct and 17 from negative direction.
Uses MC_TouchProbe and needed to add fbHalt.
Routine drives the motor in a certain direction and latches the position
of the home signal when detected. Halts the motor and then compares
current position to latched position to set the position using fbHomeDirect.